Split algorithm-independent static-method tests into their own class
extractRawGS2Header, decodeSaslname, rejectReservedMandatoryExtension, and hasValidBase64Length tests don't depend on the hash algorithm, but ran three times each via AbstractScramSaslServerTest for no added coverage. Moved to a new ScramSaslServerStaticMethodsTest that runs once.
